+++ flores/__init__.py
@@ -1,9 +1,4 @@
 """Definitions for the package-level information for Flores."""
 
-name = "flores"
-description = "Yet another static site generator."
-__version_info__ = ("0", "1", "2")
+__version_info__ = ("0", "2", "0")
 __version__ = ".".join(__version_info__)
-__author_name__ = "Dimitri Kokkonis"
-__author_email__ = "kokkonisd@gmail.com"
-__author__ = f"{__author_name__} ({__author_email__})"