+++ tests/generator/test_generate_site_with_user_data_pages.py
@@ -3,7 +3,7 @@ import re
 
 import pytest
 
-from flores.exceptions import FileOrDirNotFoundError, TemplateError
+from flores.exceptions import FileOrDirNotFoundError, FloresError, TemplateError
 from flores.generator import Generator
 
 
@@ -64,6 +64,26 @@ def test_generate_user_data_pages_with_missing_template(test_data_dir: str) -> N
         generator.build()
 
 
+def test_generate_user_data_pages_with_permalinks(test_data_dir: str) -> None:
+    """Attempt to generate a site with user data pages that contain permalinks.
+
+    Since permalinks are not allowed for user data pages, we expect them to trigger an
+    error that's reported along with the name of the offending page file.
+    """
+    generator = Generator(
+        os.path.join(test_data_dir, "user_data_pages_with_permalinks")
+    )
+
+    with pytest.raises(
+        FloresError,
+        match=re.escape(
+            f"{os.path.join(generator.project_dir, '_mypages', 'hello.md')}: "
+            "Permalinks are not allowed for user data pages."
+        ),
+    ):
+        generator.build()
+
+
 def test_generate_site_with_user_data_pages(test_data_dir: str) -> None:
     """Generate a site containing custom user data pages.
 
@@ -78,8 +98,16 @@ def test_generate_site_with_user_data_pages(test_data_dir: str) -> None:
         os.path.join(generator.build_dir, "mypages"),
         os.path.join(generator.build_dir, "myotherpages"),
         os.path.join(generator.build_dir, "mypages", "my-first-page.html"),
+        os.path.join(generator.build_dir, "mypages", "my-first-page"),
+        os.path.join(generator.build_dir, "mypages", "my-first-page", "index.html"),
         os.path.join(generator.build_dir, "mypages", "my-second-page.html"),
+        os.path.join(generator.build_dir, "mypages", "my-second-page"),
+        os.path.join(generator.build_dir, "mypages", "my-second-page", "index.html"),
         os.path.join(generator.build_dir, "myotherpages", "my-third-page.html"),
+        os.path.join(generator.build_dir, "myotherpages", "my-third-page"),
+        os.path.join(
+            generator.build_dir, "myotherpages", "my-third-page", "index.html"
+        ),
     ]
 
     actual_elements = []
@@ -125,14 +153,33 @@ def test_generate_site_with_user_data_pages(test_data_dir: str) -> None:
     ) as file:
         actual_first_page = file.read()
     with open(
+        os.path.join(generator.build_dir, "mypages", "my-first-page", "index.html"), "r"
+    ) as file:
+        actual_first_page_with_slash = file.read()
+    with open(
         os.path.join(generator.build_dir, "mypages", "my-second-page.html"), "r"
     ) as file:
         actual_second_page = file.read()
     with open(
+        os.path.join(generator.build_dir, "mypages", "my-second-page", "index.html"),
+        "r",
+    ) as file:
+        actual_second_page_with_slash = file.read()
+    with open(
         os.path.join(generator.build_dir, "myotherpages", "my-third-page.html"), "r"
     ) as file:
         actual_third_page = file.read()
+    with open(
+        os.path.join(
+            generator.build_dir, "myotherpages", "my-third-page", "index.html"
+        ),
+        "r",
+    ) as file:
+        actual_third_page_with_slash = file.read()
 
     assert actual_first_page == expected_first_page
+    assert actual_first_page_with_slash == expected_first_page
     assert actual_second_page == expected_second_page
+    assert actual_second_page_with_slash == expected_second_page
     assert actual_third_page == expected_third_page
+    assert actual_third_page_with_slash == expected_third_page